refactor: 사용자 입력 부분 view 폴더로 이동
