+++ src/view/User.js
@@ -1,7 +1,7 @@
 import { Console } from '@woowacourse/mission-utils';
-import { ERROR_MESSAGE, INPUT_MESSAGE } from './constants/message.js';
-import { validation } from './utils/validation.js';
-import ValidateError from './error/ValidateError.js';
+import { ERROR_MESSAGE, INPUT_MESSAGE } from '../constants/message.js';
+import { validation } from '../utils/validation.js';
+import ValidateError from '../error/ValidateError.js';
 
 class User {
   async getLottoPurchasePrice() {
